Rest api

+++ src/main/java/com/netckracker/graph/manager/controller/NodeController.java
@@ -10,6 +10,7 @@ import com.netckracker.graph.manager.modelDto.GraphDto;
 import com.netckracker.graph.manager.modelDto.NodeDto;
 import com.netckracker.graph.manager.modelDto.ResourceDto;
 import com.netckracker.graph.manager.service.NodeService;
+import com.netckracker.graph.manager.service.ReceipeService;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -30,57 +31,104 @@ import org.springframework.web.bind.annotation.RestController;
 public class NodeController {
     @Autowired
     private NodeService nodeService;
+    @Autowired
+    private ReceipeService receipeService;
     
     @RequestMapping(value = "/node/addnode", method = RequestMethod.POST, 
             consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE, headers = "Accept=application/json")
     public ResponseEntity<String> addNode(@RequestParam String receipeId, @RequestParam String userId){
         String nodeId=nodeService.createNode(receipeId, userId);
-        return new ResponseEntity<>(nodeId, HttpStatus.OK);
+        if (nodeId==null)
+        {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        else return new ResponseEntity<>(nodeId, HttpStatus.OK);
     }
     
     @RequestMapping(value = "/node/deletenode/{nodeId}", method = RequestMethod.DELETE)
     public ResponseEntity<Void> deleteNode(@PathVariable String nodeId){
-        nodeService.deleteNode(nodeId);
-        return new ResponseEntity<>( HttpStatus.OK);
+        if (nodeService.isNodeExcist(nodeId)==true)
+        {
+            nodeService.deleteNode(nodeId);
+            return new ResponseEntity<>( HttpStatus.OK);
+        }
+        else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
     
     @RequestMapping(value = "/node/deleteedge", method = RequestMethod.DELETE)
     public ResponseEntity<Void> deleteEdge(@RequestParam ("startNodeId") String startNodeId, @RequestParam ("endNodeId") String endNodeId){
-    nodeService.deleteEdge(startNodeId, endNodeId);
-    return new ResponseEntity<>( HttpStatus.OK);
-}
+        if (nodeService.isNodeExcist(startNodeId)==true&&nodeService.isNodeExcist(endNodeId)==true)
+        {
+            nodeService.deleteEdge(startNodeId, endNodeId);
+            return new ResponseEntity<>( HttpStatus.OK);
+        }
+        else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
     
     @RequestMapping(value = "/node/addnodedescription/{nodeId}", method = RequestMethod.POST, 
             consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE, headers = "Accept=application/json")
     public ResponseEntity<Void> addNodeDescription(@PathVariable String nodeId, @RequestParam String description){
-        nodeService.addNodeDescription(nodeId, description);
-        return new ResponseEntity<>(HttpStatus.OK);
+        if (nodeService.isNodeExcist(nodeId)==true)
+        {
+            nodeService.addNodeDescription(nodeId, description);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
+    
+        @RequestMapping(value = "/node/addnodedelabel/{nodeId}", method = RequestMethod.POST, 
+            consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE, headers = "Accept=application/json")
+    public ResponseEntity<Void> addNodeLabel(@PathVariable String nodeId, @RequestParam String label){
+        if (nodeService.isNodeExcist(nodeId)==true)
+        {
+            nodeService.addNodeLabel(nodeId, label);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+    
     @RequestMapping(value = "/node/addnodepicture/{nodeId}", method = RequestMethod.POST, 
             consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE, headers = "Accept=application/json")
     public ResponseEntity<Void> addNodePicture(@PathVariable String nodeId, @RequestParam String pictureId){
-        nodeService.addNodePicture(nodeId, pictureId);
-        return new ResponseEntity<>(HttpStatus.OK);
+        if (nodeService.isNodeExcist(nodeId)==true)
+        {
+            nodeService.addNodePicture(nodeId, pictureId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
+    
     @RequestMapping(value = "/node/addedge", method = RequestMethod.POST, 
             consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE, headers = "Accept=application/json")
     public ResponseEntity<Void> addEdge(@RequestParam String startNodeId, @RequestParam String endNodeId){
-        nodeService.createEdge(startNodeId, endNodeId);
-        return new ResponseEntity<> (HttpStatus.OK);
+        if (nodeService.isNodeExcist(startNodeId)==true&&nodeService.isNodeExcist(endNodeId)==true)
+        {
+            nodeService.createEdge(startNodeId, endNodeId);
+            return new ResponseEntity<> (HttpStatus.OK);
+        }
+        else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
     
     @RequestMapping(value = "/node/addinputresources/{nodeId}", method = RequestMethod.POST,
             consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE, headers = "Accept=application/json")
     public ResponseEntity<Void> addInputResources(@PathVariable String nodeId, @RequestBody  List<ResourceDto> resources){
-        nodeService.addInputOrOutputResourcesToNode(nodeId, resources, "input");
-        return new ResponseEntity<>(HttpStatus.OK);
+        if (nodeService.isNodeExcist(nodeId)==true)
+        {
+            nodeService.addInputOrOutputResourcesToNode(nodeId, resources, "input");
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
     
     @RequestMapping(value = "/node/addoutputresources/{nodeId}", method = RequestMethod.POST,
             consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE, headers = "Accept=application/json")
     public ResponseEntity<Void> addOutputResources(@PathVariable String nodeId, @RequestBody  List<ResourceDto> resources){
-        nodeService.addInputOrOutputResourcesToNode(nodeId, resources, "output");
-     return new ResponseEntity<>(HttpStatus.OK);
+        if (nodeService.isNodeExcist(nodeId)==true)
+        {
+            nodeService.addInputOrOutputResourcesToNode(nodeId, resources, "output");
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
     
     @RequestMapping(value = "/node/getinputresources/{nodeId}", method = RequestMethod.GET,
@@ -109,8 +157,42 @@ public class NodeController {
     @RequestMapping(value = "/graph/getgraph", method = RequestMethod.GET, 
             consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE, headers = "Accept=application/json")
     public ResponseEntity<?> getGraph( @RequestParam ("userId") String userId, @RequestParam ("receipeId") String receipeId){
-                GraphDto graph=nodeService.getReceipeGraph(receipeId, userId);
-            return new ResponseEntity<>(graph, HttpStatus.OK);        
+        if (receipeService.isVersionCompleted(receipeId)==true)
+        {
+            GraphDto graph=nodeService.getReceipeGraph(receipeId, userId);
+            if (graph==null)
+            {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+            else return new ResponseEntity<>(graph, HttpStatus.OK); 
+        }
+        else return new ResponseEntity<>(HttpStatus.NOT_FOUND);     
+    }
+    @RequestMapping(value = "/graph/getparallelgraph", method = RequestMethod.GET, 
+            consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE, headers = "Accept=application/json")
+    public ResponseEntity<?> getParallelGraph( @RequestParam ("userId") String userId, @RequestParam ("receipeId") String receipeId){
+        if (receipeService.isVersionCompleted(receipeId)==true)
+        {
+            GraphDto graph=nodeService.getReceipeParallelGraph(receipeId, userId);
+            if (graph==null)
+            {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+            else return new ResponseEntity<>(graph, HttpStatus.OK);  
+        }
+        else return new ResponseEntity<>(HttpStatus.NOT_FOUND);  
+    }
+    
+    @RequestMapping(value = "/graph/getnotcompletedgraph", method = RequestMethod.GET, 
+            consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE, headers = "Accept=application/json")
+    public ResponseEntity<?> getNotCompletedGraph( @RequestParam ("userId") String userId, @RequestParam ("receipeId") String receipeId){
+        
+            GraphDto graph=nodeService.getReceipeGraph(receipeId, userId);
+            if (graph==null)
+            {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+            else return new ResponseEntity<>(graph, HttpStatus.OK);  
     }
     
     @RequestMapping(value = "/graph/gettestgraph", method = RequestMethod.GET)
@@ -118,5 +200,7 @@ public class NodeController {
                 GraphDto graph=nodeService.getReceipeTestGraph(receipeId, userId);
             return new ResponseEntity<>(graph, HttpStatus.OK);        
     }  
+    
+    
  
 }
